requirements.txt
requests
sseclient-py
kafka-python
pyspark
fastapi
uvicorn
duckdb
pandas
pyarrow
delta-spark
python-dotenv